+++ lib/teamleader/api.rb
@@ -41,8 +41,9 @@ module Teamleader
       request "/addCompany.php", params
     end
 
-    def get_company(id)
-      request "/getCompany.php", {:company_id => id}
+    def get_company(params={})
+      raise "company_id is required" if params[:company_id].nil?
+      request "/getCompany.php", params
     end
 
     def get_companies(params={})
@@ -52,20 +53,65 @@ module Teamleader
     end
 
     def add_invoice(params={})
-      raise "contact_or_company" if params[:contact_or_company].nil?
-      raise "contact_or_company_id" if params[:contact_or_company_id].nil?
-      raise "sys_department_id" if params[:sys_department_id].nil?
+      raise "contact_or_company is required" if params[:contact_or_company].nil?
+      raise "contact_or_company_id is required" if params[:contact_or_company_id].nil?
+      raise "sys_department_id is required" if params[:sys_department_id].nil?
       request "/addInvoice.php", params    
     end
 
-    def get_invoice(id)
-      request "/getInvoice.php", {:invoice_id => id}
+    def update_invoice_payment_status(params={})
+      raise "invoice_id is required" if params[:invoice_id].nil?
+      raise "status is required" if params[:status].nil?
+      raise "status must be 'paid' or 'not_paid'" unless ['paid', 'not_paid'].include?(params[:status])
+      request "/setInvoicePaymentStatus.php", params    
+    end
+
+    def get_invoice(params={})
+      raise "invoice_id is required" if params[:invoice_id].nil?
+      request "/getInvoice.php", params
+    end
+
+    def download_invoice_pdf(params={})
+      raise "invoice_id is required" if params[:invoice_id].nil?
+      request "/downloadInvoicePDF.php", params
     end
 
     def get_invoices(params={})
       raise "date_from is required" if params[:date_from].nil?
       raise "date_to is required" if params[:date_to].nil?
       request "/getInvoices.php", params
+    end
+
+    def add_product(params={})
+      raise "name is required" if params[:name].nil?
+      raise "price is required" if params[:price].nil?
+      request "/addProduct.php", params
+    end
+
+    def update_product(params={})
+      raise "product_id is required" if params[:product_id].nil?
+      request "/updateProduct.php", params
+    end
+
+    def delete_product(params={})
+      raise "product_id is required" if params[:product_id].nil?
+      request "/deleteProduct.php", params
+    end
+
+    def get_product(params={})
+      raise "product_id is required" if params[:product_id].nil?
+      request "/getProduct.php", params
+    end
+
+    def get_product(params={})
+      raise "product_id is required" if params[:product_id].nil?
+      request "/getProduct.php", params
+    end
+
+    def get_products(params={})
+      raise "amount is required" if params[:amount].nil?
+      raise "pageno is required" if params[:pageno].nil?
+      request "/getProducts.php", params
     end
 
     private