+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 2017-05-12
+### Added
+- Method `link_contact_to_company`
+- Methods for Products (add, update, delete, get one, get all)
+- Methods for Invoice (add, update payment status, get PDF, get one, get all)
+- Update `get_company` to accept an Hash instead of an id as parameter
+
 ## [0.1.1] - 2017-05-12
 ### Changed
 - Mark `json` dependency version as `>= 1`